+++ passes/R014/testdata/src/a/main_v2.go
@@ -1,6 +1,11 @@
 package a
 
-import "github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+import (
+	"context"
+
+	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+)
 
 // Comment ignored
 
@@ -10,8 +15,18 @@ func commentIgnoreFirstParameter_v2(invalid *schema.ResourceData, meta interface
 }
 
 //lintignore:R014
+func commentIgnoreFirstParameter_v2_context(ctx context.Context, invalid *schema.ResourceData, meta interface{}) diag.Diagnostics {
+	return diag.Diagnostics{}
+}
+
+//lintignore:R014
 func commentIgnoreSecondParameter_v2(d *schema.ResourceData, invalid interface{}) error {
 	return nil
+}
+
+//lintignore:R014
+func commentIgnoreSecondParameter_v2_context(ctx context.Context, d *schema.ResourceData, invalid interface{}) diag.Diagnostics {
+	return diag.Diagnostics{}
 }
 
 // Failing
@@ -22,9 +37,21 @@ func failingAnonymousFirstParameter_v2() {
 	}
 }
 
+func failingAnonymousFirstParameter_v2_context() {
+	_ = func(ctx context.Context, invalid *schema.ResourceData, meta interface{}) diag.Diagnostics { // want "\\*schema.ResourceData parameter of CreateContextFunc, ReadContextFunc, UpdateContextFunc, or DeleteContextFunc should be named d"
+		return diag.Diagnostics{}
+	}
+}
+
 func failingAnonymousSecondParameter_v2() {
 	_ = func(d *schema.ResourceData, invalid interface{}) error { // want "interface\\{\\} parameter of CreateFunc, ReadFunc, UpdateFunc, or DeleteFunc should be named meta"
 		return nil
+	}
+}
+
+func failingAnonymousSecondParameter_v2_context() {
+	_ = func(ctx context.Context, d *schema.ResourceData, invalid interface{}) diag.Diagnostics { // want "interface\\{\\} parameter of CreateContextFunc, ReadContextFunc, UpdateContextFunc, or DeleteContextFunc should be named meta"
+		return diag.Diagnostics{}
 	}
 }
 
@@ -32,8 +59,16 @@ func failingFirstParameter_v2(invalid *schema.ResourceData, meta interface{}) er
 	return nil
 }
 
+func failingFirstParameter_v2_context(ctx context.Context, invalid *schema.ResourceData, meta interface{}) diag.Diagnostics { // want "\\*schema.ResourceData parameter of CreateContextFunc, ReadContextFunc, UpdateContextFunc, or DeleteContextFunc should be named d"
+	return diag.Diagnostics{}
+}
+
 func failingSecondParameter_v2(d *schema.ResourceData, invalid interface{}) error { // want "interface\\{\\} parameter of CreateFunc, ReadFunc, UpdateFunc, or DeleteFunc should be named meta"
 	return nil
+}
+
+func failingSecondParameter_v2_context(ctx context.Context, d *schema.ResourceData, invalid interface{}) diag.Diagnostics { // want "interface\\{\\} parameter of CreateContextFunc, ReadContextFunc, UpdateContextFunc, or DeleteContextFunc should be named meta"
+	return diag.Diagnostics{}
 }
 
 // Passing
@@ -44,10 +79,32 @@ func passingAnonymous_v2() {
 	}
 }
 
+func passingAnonymous_v2_context() {
+	_ = func(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
+		return diag.Diagnostics{}
+	}
+}
+
 func passing_v2(d *schema.ResourceData, meta interface{}) error {
 	return nil
 }
 
+func passing_v2_context(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
+	return diag.Diagnostics{}
+}
+
 func passingOther_v2(diff *schema.ResourceDiff, meta interface{}) error {
 	return nil
+}
+
+func passingOther_v2_context(ctx context.Context, diff *schema.ResourceDiff, meta interface{}) diag.Diagnostics {
+	return diag.Diagnostics{}
+}
+
+func passingExtended_v2(invalidResourceData *schema.ResourceData, invalidInterface interface{}, notCRUDFunc string) error {
+	return nil
+}
+
+func passingExtended_v2_context(invalidContext context.Context, invalidResourceData *schema.ResourceData, invalidInterface interface{}, notCRUDContextFunc string) diag.Diagnostics {
+	return diag.Diagnostics{}
 }